requirements.txt
a2a-sdk[http-server]>=0.3.25
anthropic>=0.87.0
uvicorn>=0.42.0
